Use AJAX class instead
